moving wallets to root
